Added a virtual MAX Wallthermostat
First implementation of a virtual Wall thermostat as already described in Issue: https://github.com/Homegear/Homegear/issues/313

As I'm not exactly sure how homegear works I took the implementation of a virtual Homematic-BidCos device for example and tried to adopt it. This surely needs some more work to do.